Update README.md
Axon, a classic car retailer, faces challenges managing and analyzing sales data due to the absence of a centralized system. To address this, they've chosen to implement a Business Intelligence (BI) solution. Microsoft Power BI and SQL are top contenders for this initiative. The goal is to deploy an efficient BI solution, empowering Axon to make data-driven decisions and enhance their competitive edge in the classic car retail market.